refactor: extract shared telemetry wrapper from message and app_mention handlers
Eliminate duplicate try/catch/finally telemetry blocks by delegating to
handleInteractionWithTelemetry in interaction-telemetry.js. Both handlers
now use claimResponseId, markRateLimited, and markInteractionRecorded
helpers instead of managing shared mutable state directly.
